requirements.txt
asgiref==3.8.1
attrs==23.2.0
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.16.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.0.3
djangorestframework==3.15.1
docx2pdf==0.1.8
ffmpeg-python==0.2.0
future==1.0.0
mysqlclient==2.2.4
pillow==10.3.0
pycparser==2.22
pyheif==0.7.1
pyheif-pillow-opener==0.2.0
PyPDF2==3.0.1
sqlparse==0.4.4
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.10.0
webencodings==0.5.1
